Move TakeControlDialog into TakeControl folder inside __buttons__
