+++ src/web/components/__buttons__/DataOfffloadAllButton/DataOffloadAllButton.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 
-import TakeControlDialog from "../../TakeControlDialog/TakeControlDialog";
+import TakeControlDialog from "../TakeControl/TakeControlDialog/TakeControlDialog";
 import { DataOffloadAllDialog } from "./DataOffloadAllDialog";
 import { DisabledCodes } from "../disabled-codes";
