+++ src/web/components/__buttons__/DataOffloadButton/DataOffloadButton.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 
-import TakeControlDialog from "../../TakeControlDialog/TakeControlDialog";
+import TakeControlDialog from "../TakeControl/TakeControlDialog/TakeControlDialog";
 import { DataOffloadDialog } from "./DataOffloadDialog";
 import { DisabledCodes } from "../disabled-codes";
