+++ src/web/components/__buttons__/NextTaskButton/NextTaskButton.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 
-import TakeControlDialog from "../../TakeControlDialog/TakeControlDialog";
+import TakeControlDialog from "../TakeControl/TakeControlDialog/TakeControlDialog";
 import { NextTaskDialog } from "./NextTaskDialog";
 import { DisabledCodes } from "../disabled-codes";
