+++ src/web/components/__buttons__/StartMissionButton/StartMissionButton.tsx
@@ -2,7 +2,7 @@ import { useContext, useState } from "react";
 import { JaiaDispatchContext } from "../../../context/JaiaContext";
 import { JaiaActions } from "../../../context/jaia-actions";
 
-import TakeControlDialog from "../../TakeControlDialog/TakeControlDialog";
+import TakeControlDialog from "../TakeControl/TakeControlDialog/TakeControlDialog";
 import { StartMissionDialog } from "./StartMissionDialog";
 import { DisabledCodes } from "../disabled-codes";
