+++ src/web/components/__buttons__/SystemButton/SystemButton.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 
-import TakeControlDialog from "../../TakeControlDialog/TakeControlDialog";
+import TakeControlDialog from "../TakeControl/TakeControlDialog/TakeControlDialog";
 import { SystemDialog } from "./SystemDialog";
 import { DisabledCodes } from "../disabled-codes";
